Rename premium content constants file
